refactored nlp models to have helper functions in their own module
